+++ src/java/model/Payment.java
@@ -12,7 +12,7 @@ import java.util.Date;
  */
 public class Payment {
     public int id;
-    public int amount;
+    public float amount;
     public Date date = new Date();
     public String username;
     public String typeOfPayment;
@@ -28,7 +28,7 @@ public class Payment {
     
     public String getUsername(){return this.username;}
     
-    public int getAmount() {return this.amount;}
+    public float getAmount() {return this.amount;}
     
     public String getTypeOfPayment(){return this.typeOfPayment;}
         
@@ -36,7 +36,7 @@ public class Payment {
     
     public void setUsername(String username){this.username = username;}
     
-    public void setAmount(int amount){this.amount = amount;}
+    public void setAmount(float amount){this.amount = amount;}
         
     public void setTypeOfPayment(String typeOfPayment){this.typeOfPayment = typeOfPayment;}
